adding float handeling to avg and sum

+++ lifetracker-ui/src/components/ActivityPage/ActivityPage.jsx
@@ -22,8 +22,8 @@ export default function ActivityPage({appState, isLoggedIn, exerciseArray, setEx
               const exercises = response.data.exercises;
               const exerciseTime = response.data.exerciseTime;
               const exerciseAvg = response.data.exerciseAvg;
-              setExerciseAvg(exerciseAvg === null ? 0.00 : (parseInt(exerciseAvg).toFixed(1)));
-              setExerciseTime(exerciseTime === null ? 0.00 : (parseInt(exerciseTime).toFixed(1)));
+              setExerciseAvg(exerciseAvg === null ? 0.00 : (parseFloat(exerciseAvg).toFixed(1)));
+              setExerciseTime(exerciseTime === null ? 0.00 : (parseFloat(exerciseTime).toFixed(1)));
             })
             .catch((error) => {
               console.log("Error with axios:", error); // Debugging console.log
@@ -44,8 +44,8 @@ export default function ActivityPage({appState, isLoggedIn, exerciseArray, setEx
             console.log("calories", response.data)
               const nutritionCalories = response.data.nutritionCalories;
               const nutritionMax = response.data.nutritionMax;
-              setNutritionCalories(nutritionCalories === null ? 0.00 : (parseInt(nutritionCalories).toFixed(1)))
-              setNutritionMax(nutritionMax === null? 0.00 : (parseInt(nutritionMax).toFixed(1)))
+              setNutritionCalories(nutritionCalories === null ? 0.00 : (parseFloat(nutritionCalories).toFixed(1)))
+              setNutritionMax(nutritionMax === null? 0.00 : (parseFloat(nutritionMax).toFixed(1)))
             })
             .catch((error) => {
               console.log("Error with axios:", error); // Debugging console.log
